Changes:
- fixed a bug where the alpine test container was getting pulled on every test (~10 seconds per test)
- initialize reusable objects once at beginning of test (~22 seconds per test)
- tear down reusable objects once at the end of the tests (~6 seconds per test)

Total time saved: 38 seconds per test (total 8 minutes across 13 tests)

Notes:
Half the test time is used by podman for testing pushing containers. Even when the user doesn't have permissions to push containers, this step still takes 10 seconds and runs 3 times per test. I would like to find a way to further optimize this, but there doesn't seem to be any way around the fact that podman is just really slow.